rename the postModels.js file as postModel.js
