+++ src/api/v1/middleware/existenceValidators.js
@@ -1,5 +1,5 @@
 import { existenceValidatorBuilder } from '../../helpers/existenceValidatorBuilder.js'
-import { existsById } from '../models/postModels.js'
+import { existsById } from '../models/postModel.js'
 
 const postExists = existenceValidatorBuilder(existsById, 'La publicación no existe')
